Fix up tap target heights.
The tap targets are really big and are getting really close to
the seek bar. Reduce the size a little to avoid mistriggers.

Also, the default dimens file was missing a bunch of resources
which could cause the application to crash because it wouldn't
find a resource at some resolutions. Put in placeholder numbers
for those.

Bug: 36191136
Test: Looked at the layout bounds with show layout bounds enabled
      and made sure they look reasonable.
Change-Id: Id55f397acb8e751a75978b100051f34e6c5eb2dd