Decouple CarDrawerActivity from media drawer classes.
The CarDrawerActivity is now only passed to the MediaDrawerController.
The controller is now completely responsible for all drawer navigations.
The various item fetchers will notify the controller via interfaces of
variious user actions (e.g. clicks).

It is possible to separate the CarDrawerActivity from the
MediaDrawerController, but this is cleaner if done from the
car-stream-ui-lib level. That is, make some kind of base
CarDrawerController.

Test: Started media app. Test playing, opening queue and navigation in
drawer.

Change-Id: Idaaedf19ec2127f9b673d4294e08901b1dd96886