Scroll to currently playing item.
We moved to showing the whole queue so make the list scroll
to the currently playing item. We need to scroll both on the
queue open and during the viewholder binding since the viewholders
for positions that don't show up on the screen are not guaranteed
to be called when the list becomes visible. By forcing the
initial scroll, we can make sure the updates while binding take
place.

Bug: 37999972
Test: 1. Skipped forward many tracks and opened the queue. Verified
         that the currently playing song was scrolled to.
      2. Stayed with the queue open and watched the list scroll when
         the track changed.

Change-Id: I424b68c0fe3a74073e53495246065ca6f0244891