Add a settings button for media apps that declare a preference activity
When the selected media app declares a preference activity by including
an APPLICATION_PREFERENCES intent-filter (see cl/212053225), then a
settings button is shown in the AppBarView that opens the app's preference
activity when pressed.

UX will probably want to use an icon and move the button elsewhere, but
this allows to test the feature.

Test: manual
Change-Id: Iaf50ba38f4c388ec21afe6afc64b65d24d28c202